+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 # Changelog
 
-- Postagem de item adicionada
+- Sistema de compra de item
+- Sistema de conversa com vendedor (parcialmente adicionada)
+- Pequeno re-work na database
+- Criação de database para compradores
+- Criação de mensagens via embed (estética)
+- BugFixes
+- https://youtu.be/LgA7vtoYWZE